requirements.txt
openai==1.59.9
python-dotenv==1.0.1
python-multipart==0.0.20
streamlit==1.41.1
